fix centOS 8 test failures

+++ scripts/eosio_build_centos.sh
@@ -42,6 +42,7 @@ if [[ "$(echo ${VERSION} | sed 's/ .*//g')" == 8 ]]; then
         install-package https://dl.fedoraproject.org/pub/epel/epel-release-latest-8.noarch.rpm
         group-install-package 'Development Tools'
         install-package openssl-devel
+        install-package which
 	install-package git 
 	install-package autoconf
 	install-package automake
@@ -56,6 +57,8 @@ if [[ "$(echo ${VERSION} | sed 's/ .*//g')" == 8 ]]; then
 	install-package libusbx-devel
     	install-package libcurl-devel
 	install-package patch
+        install-package vim-common
+        install-package jq
     	install-package python3
 	install-package python3-devel
 	install-package clang
@@ -63,9 +66,27 @@ if [[ "$(echo ${VERSION} | sed 's/ .*//g')" == 8 ]]; then
 	install-package llvm-static
 	install-package procps-ng
 	install-package util-linux
+        install-package sudo
 	install-package libstdc++
+        install-package dnf-plugins-core
+        sudo dnf config-manager --set-enabled PowerTools
+        install-package doxygen
+        install-package ocaml
 	install-package ncurses-compat-libs
-	ln -s /usr/lib64/libtinfo.so.6 /usr/local/lib/libtinfo.so
+        install-package nodejs
+        install-package epel-release
+
+        curl --silent --location https://dl.yarnpkg.com/rpm/yarn.repo | sudo tee /etc/yum.repos.d/yarn.repo
+        sudo rpm --import https://dl.yarnpkg.com/rpm/pubkey.gpg
+        install-package yarn
+
+        pushd ${REPO_ROOT}
+        yarn install
+        popd
+
+        if [ ! -L "/usr/local/lib/libtinfo.so" ] ; then
+	    ln -s /usr/lib64/libtinfo.so.6 /usr/local/lib/libtinfo.so
+        fi
 fi
 
 # Handle clang/compiler
